optimised model loading, but am using the gltf crate instead of russimp-ng.
this lowers down the import speed for 2 models from 10 seconds to 3 seconds. I'm not exactly sure how to optimise it, but im sure as hell that i need to implement async/await into the project. i suppose thats for another time...

this should resolve Optimise Model Loading
Fixes #41 for the timebeing